# profiler/internal/fastdelta: handle duplicate samples