--- a/profiler/internal/fastdelta/delta_map.go
+++ b/profiler/internal/fastdelta/delta_map.go
@@ -26,16 +26,22 @@ type sampleValue [maxSampleValues]int64
 func NewDeltaMap(st *stringTable, lx *locationIndex, fields []valueType) *DeltaMap {
 	return &DeltaMap{
 		h:                    Hasher{alg: murmur3.New128(), st: st, lx: lx},
-		m:                    map[Hash]sampleValue{},
+		m:                    map[Hash]combinedSampleValue{},
 		st:                   st,
 		fields:               fields,
 		computeDeltaForValue: make([]bool, 0, 4),
 	}
 }
 
+type combinedSampleValue struct {
+	old     sampleValue
+	new     sampleValue
+	written bool
+}
+
 type DeltaMap struct {
 	h  Hasher
-	m  map[Hash]sampleValue
+	m  map[Hash]combinedSampleValue
 	st *stringTable
 	// fields are the name and types of the values in a sample for which we should
 	// compute the difference.
@@ -56,6 +62,36 @@ func (dm *DeltaMap) AddSampleType(st *pproflite.SampleType) error {
 	return nil
 }
 
+func (dm *DeltaMap) UpdateSample(sample *pproflite.Sample) error {
+	if err := dm.prepare(); err != nil {
+		return err
+	}
+
+	hash, err := dm.h.Sample(sample)
+	if err != nil {
+		return err
+	}
+
+	var val sampleValue
+	n := 0
+	for i := range sample.Value {
+		if dm.computeDeltaForValue[i] {
+			val[n] = sample.Value[i]
+			n++
+		}
+	}
+
+	var c combinedSampleValue
+	old := dm.m[hash]
+	c.old = old.old
+	for i, v := range old.new {
+		val[i] += v
+	}
+	c.new = val
+	dm.m[hash] = c
+	return nil
+}
+
 // Delta updates sample.Value by looking up the previous values for this sample
 // and substracting them from the current values. The returned boolean is true
 // if the the new sample.Value contains at least one non-zero value.
@@ -69,38 +105,35 @@ func (dm *DeltaMap) Delta(sample *pproflite.Sample) (bool, error) {
 		return false, err
 	}
 
-	var val sampleValue
+	c, ok := dm.m[hash]
+	if !ok {
+		// !ok should not happen, since the prior pass visited every sample
+		return false, fmt.Errorf("found sample with unknown hash in merge pass")
+	}
+	if c.written {
+		return false, nil
+	}
+	all0 := true
 	n := 0
 	for i := range sample.Value {
 		if dm.computeDeltaForValue[i] {
-			val[n] = sample.Value[i]
+			sample.Value[i] = c.new[n] - c.old[n]
 			n++
 		}
-	}
-
-	old, ok := dm.m[hash]
-	dm.m[hash] = val // save for next time
-	if ok {
-		all0 := true
-		n := 0
-		for i := range sample.Value {
-			if dm.computeDeltaForValue[i] {
-				sample.Value[i] -= old[n]
-				n++
-			}
-			if sample.Value[i] != 0 {
-				all0 = false
-			}
-		}
-
-		if all0 {
-			// If the sample has all 0 values, we drop it
-			// this matches the behavior of Google's pprof library
-			// when merging profiles
-			return false, nil
+		if sample.Value[i] != 0 {
+			all0 = false
 		}
 	}
-	return true, nil
+
+	c.written = true
+	c.old = c.new
+	c.new = sampleValue{}
+	dm.m[hash] = c
+
+	// If the sample has all 0 values, we drop it
+	// this matches the behavior of Google's pprof library
+	// when merging profiles
+	return !all0, nil
 }
 
 func (dm *DeltaMap) prepare() error {