@nsrip-dd 2022-11-10
The fix appears to have eliminated the fidelity failures in the reliability environment:

<img width="842" alt="Screen Shot 2022-11-10 at 9 01 22 AM" src="https://user-images.githubusercontent.com/97066770/201111903-06cca90d-bd61-4373-b041-2927ff2dd3dc.png">