@nsrip-dd 2022-11-11
Turns out this is expected behavior and it's right to account for it. I've written it up here: https://gist.github.com/nsrip-dd/df0b6702a7b16a9612f52cd36908e196 TL;DR: when encoding the heap profile, Go will remove any leaf runtime function calls so that the call stack terminates at a non-runtime function. Such calls would be tracked as separate allocations internally, but when truncated show the same call stack in the encoded profile.